requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
plotly>=5.15.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0 